Create a metric for author's knowledge of code base
This is a very rudimentary initial work, mostly to get familiar
with writing Chrome plugins. It does, however, add an actual metric,
which is the percentage of merged PRs the author of this PR has also authored.

This gives a rough gauge of how familiar the developer is with the
code base. More familiar is a lower risk factor.