Provide initial support for per-diff risks.
This is an initial starting point for providing per-diff
risk assessments, which are different than the overall assessments.

This PR adds support for creating the UI and an initial risk assessment:
max complexity in the method. There will definitely need to be some
refactoring as this progresses.

This PR also makes mild changes around CSS values and names to acknowledge
the similarities across different types of assessment tables.