Add "file volatility" and "author volatility risk factor to diffs.
This necessitated methods to traverse git responses,
which I've placed in git_helper.js. This will become the home
for all git-specific logic moving forward, but this function is the first
addition.

In addition, updated the README and modified the CSS to
give a consistent look to all prisk tables