Allow filtering of files from the risk assessments.
This PR allows the system to have a notion of
excluded files: ones unlikely to cause a real problem.

For instance, ReleaseNotes.md might get updated with every PR,
but an issue with that is unlikely to cause a non-trivial defect.

This required a bit of refactoring and bug fixing, which is included in the PR