Replace multi-row table text with icons
For each diff risk assessment, remove the multi-row, 2-column
table and present a much more compressed view using
a single row of color-coded icons. Each icon also presents
the old text form in the alt text.

As part of this, create a new ui.js file that is the home
(or will be the home for) functions specific to the UI.