Allow optimizations on commit queries.
In the case of determining a file's youth, older files
would take longer because the code has to find the earliest
commit for the file. However, we know that if _any_ commit
is older than the "low-risk" number, we can stop right then.

This commit thus adds the ability to pass a "shouldContinue"
function to commit-fetching code. If the result is true, then
the fetching code will continue. If it's false, the results
as they exist will be returned.

Now when investigating a file, the file youth query will
stop if any commit including the file is older than the "low-risk" date.