Fix bugs wherein the UI was not updated when clicking on Pull Requests or Files.
This was caused by the fact that these pages are loaded via XHR. So this commit
adds a background script that checks for xhr requests and tells prisk
to update if it needs to. This commit also adds a "did I already run" check
to prisk to make sure xhr events do not spam the UI with divs.